separating class from main file
