+++ on-boarding/task-2.dart
@@ -1,117 +1,7 @@
 import 'dart:io';
+import 'task_manager.dart';
+import 'task.dart';
 
-class Task {
-  String? title;
-  String? description;
-  DateTime dueDate;
-  bool completed;
-
-  Task(this.title, this.description, this.dueDate, {this.completed = false});
-
-  void markAsCompleted() {
-    completed = true;
-  }
-
-  void markAsPending() {
-    completed = false;
-  }
-
-  @override
-  String toString() {
-    String statusEmoji = completed ? "✅" : "🕒";
-    return '📌 Title: $title\n📝 Description: $description\n⏰ Due Date: $dueDate\n$statusEmoji Status: ${completed ? "Completed" : "Pending"}\n';
-  }
-}
-
-class TaskManager {
-  List<Task> tasks = [];
-
-  void addTask(Task task) {
-    tasks.add(task);
-  }
-
-  void viewAllTasks() {
-    if (tasks.isEmpty) {
-      print("No tasks found. 📭");
-      return;
-    }
-    for (int i = 0; i < tasks.length; i++) {
-      print('Task ${i + 1}:');
-      print(tasks[i]);
-    }
-  }
-
-  void viewCompletedTasks() {
-    final completedTasks = tasks.where((task) => task.completed).toList();
-    if (completedTasks.isEmpty) {
-      print("No completed tasks found. ✅");
-      return;
-    }
-    for (int i = 0; i < completedTasks.length; i++) {
-      print('Completed Task ${i + 1}:');
-      print(completedTasks[i]);
-    }
-  }
-
-  void viewPendingTasks() {
-    final pendingTasks = tasks.where((task) => !task.completed).toList();
-    if (pendingTasks.isEmpty) {
-      print("No pending tasks found. 🕒");
-      return;
-    }
-    for (int i = 0; i < pendingTasks.length; i++) {
-      print('Pending Task ${i + 1}:');
-      print(pendingTasks[i]);
-    }
-  }
-
-  void editTask(int index) {
-    if (index >= 0 && index < tasks.length) {
-      Task task = tasks[index];
-      print("Current Task Details:");
-      print(task);
-
-      stdout.write("Enter new title or leave empty to keep current: ");
-      String? newTitle = stdin.readLineSync();
-      if (newTitle != null) {
-        task.title = newTitle;
-      }
-
-      stdout.write("Enter new description or leave empty to keep current: ");
-      String? newDescription = stdin.readLineSync();
-      if (newDescription != null) {
-        task.description = newDescription;
-      }
-
-      stdout.write("Enter new due date or leave empty to keep current: ");
-      String? newDueDate = stdin.readLineSync();
-      if (newDueDate != null) {
-        task.dueDate = DateTime.parse(newDueDate);
-      }
-
-      stdout.write("Is the task completed? (yes/no): ");
-      String? completedStatus = stdin.readLineSync()?.toLowerCase()?? "no";
-      if (completedStatus == 'yes') {
-        task.markAsCompleted();
-      } else {
-        task.markAsPending();
-      }
-
-      print("Task updated successfully.");
-    } else {
-      print("Invalid task index.");
-    }
-  }
-
-  void deleteTask(int index) {
-    if (index >= 0 && index < tasks.length) {
-      tasks.removeAt(index);
-      print("Task deleted successfully.");
-    } else {
-      print("Invalid task index.");
-    }
-  }
-}
 
 void main() {
   TaskManager taskManager = TaskManager();
@@ -132,10 +22,10 @@ void main() {
     switch (choice) {
       case '1':
         stdout.write("Enter task title: ");
-        String? title = stdin.readLineSync();
+        String title = stdin.readLineSync() ?? "";
 
         stdout.write("Enter task description: ");
-        String? description = stdin.readLineSync();
+        String description = stdin.readLineSync()??"";
 
         stdout.write("Enter task due date (YYYY-MM-DD): ");
         String? dueDateInput = stdin.readLineSync();
